requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.5
djangorestframework==3.12.2
lxml==4.6.2
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
